requirements.txt
langchain_community
langchain_core
langchain_text_splitters
langchain_google_genai
streamlit
pypdf
python-docx
docx2txt
unstructured
pinecone
python-dotenv
google-generativeai
python-pptx
pandas
markdown
fastapi
pydantic
python-multipart
groq